Implement bookmark functionality in BusinessHomeScreen
This commit introduces the ability to bookmark internships within the BusinessHomeScreen.

The `BusinessViewModel` has been updated to include a `bookmarkInternshipUseCase` and a `bookmarkInternship` function. This function handles the logic of marking or unmarking an internship as bookmarked and then refreshes the list of business internships.

A `TODO` comment has been removed from `BusinessHomeScreen.kt` as the bookmarking functionality is now implemented.